Add files via upload
This project is one of the many coming which show my skills. To run it, install Eclipse for Java EE and MySQL database. The mysql-script text file(on root directory) has the sql script to create database, user and a table. Maven will take care of the rest. I am available on chapfuwad@gmail.com for any question.